Updated Timing
Time is now calculated using the high resolution clock and is sampled on a microsecond scale. It is now propagated through the system as a float.
Physics is limited to run as close to 60 times per second as possible.

+++ PA11/include/camera.h
@@ -13,7 +13,7 @@ public:
 	Camera(Application* engine);
 	~Camera();
 	bool Initialize(int w, int h);
-	void Update(unsigned int dt);
+	void Update(float dt);
 	glm::mat4 GetProjection() { return projection; }
 	glm::mat4 GetView() { return view; }
