+++ PA11/include/light.h
@@ -18,7 +18,7 @@ public:
 	Light() : Light(Type::Disabled) {}
 	Light(Type type, size_t id = 0) : type(type), id(id) {}
 	void Render(Shader* boundShader) override;
-	void Update(unsigned int dt) override;
+	void Update(float dt) override;
 
 	// Light color setting
 	void setAmbient(glm::vec4 color) {lightAmbient = color;}
